+++ source/org/apache/commons/net/ftp/parser/NTFTPEntryParser.java
@@ -136,7 +136,7 @@ public class NTFTPEntryParser extends RegexFTPFileEntryParserImpl {
 				f.attributes.setType(Path.FILE_TYPE);
 				if(null != size) {
 					try {
-						f.attributes.setSize(Long.parseLong(size));
+						f.attributes.setSize(Double.parseDouble(size));
 					}
 					catch(NumberFormatException e) {
 						// intentionally do nothing
