+++ source/org/apache/commons/net/ftp/parser/OS400FTPEntryParser.java
@@ -65,7 +65,7 @@ public class OS400FTPEntryParser extends RegexFTPFileEntryParserImpl {
 			f.attributes.setOwner(usr);
 
 			try {
-				f.attributes.setSize(Long.parseLong(filesize));
+				f.attributes.setSize(Double.parseDouble(filesize));
 			}
 			catch(NumberFormatException e) {
 				// intentionally do nothing
