+++ source/org/apache/commons/net/ftp/parser/UnixFTPEntryParser.java
@@ -142,7 +142,7 @@ public class UnixFTPEntryParser extends RegexFTPFileEntryParserImpl {
 			f.attributes.setGroup(grp);
 
 			try {
-				f.attributes.setSize(Long.parseLong(filesize));
+				f.attributes.setSize(Double.parseDouble(filesize));
 			}
 			catch(NumberFormatException e) {
 				// intentionally do nothing
